タグ入力部分をコンポーネント化
