+++ app/javascript/user_tags.js
@@ -6,8 +6,12 @@ document.addEventListener('DOMContentLoaded', () => {
   if (element) {
     const tagsInitialValue = element.getAttribute('data-tags-initial-value')
     const tagsParamName = element.getAttribute('data-tags-param-name')
+    const tagsInput = element.getAttribute('data-tags-input') === 'on'
+    const tagsInputId = element.getAttribute('data-tags-input-id')
+    const tagsEditable = element.getAttribute('data-tags-editable') === 'true'
     const userId = element.getAttribute('data-user-id')
-    const editable = element.getAttribute('editable') === 'true'
+    const currentUserId = element.getAttribute('data-current-user-id')
+    const adminLogin = element.getAttribute('data-admin-login')
 
     new Vue({
       render: h => h(UserTags, {
@@ -15,7 +19,11 @@ document.addEventListener('DOMContentLoaded', () => {
           tagsInitialValue: tagsInitialValue,
           tagsParamName: tagsParamName,
           userId: userId,
-          editable: editable
+          tagsEditable: tagsEditable,
+          tagsInput: tagsInput,
+          tagsInputId: tagsInputId,
+          currentUserId: currentUserId,
+          adminLogin: adminLogin
         }
       })
     }).$mount('#js-user-tags')
